renamed MoviesListPage to MoviesList
